Add directories configuration of index: Thinking in code.
It's just not very nice. See notes in code:

Fields need adding, and reconfiguring, when a bundle is added.
A Search API field definition includes the index.

It feels like it would be cleaner to do the configuration in
the config manager, but simpler to do it in the finder plugin.