Allow plugin to set index template; Move Finder Bundle Config Form to service, test; clean-up.
= Allow plugin to set index template

Rather than adding all fields, or other settings to a very basic
template, a plugin can now add their own config/templates yml.

= Move Finder Bundle Config Form to service

Allowing reuse of the form for more entity types.
It might also make it possible for a 'Content Facets for Finders' module
to decorate the form service.

There's a functionl test that starts to check this form.

It's not yet set up that a node channel could have another entity entry.

For the switching the db backend I've investigated both adding events,
or decorating the form service; and neither really fit. Currently
wondering if the plug-ins themselves can't have some configuration page
(not necessarily config, as this would be configuring the indexes, but
maybe other things too). Might work with permissions better to: form and
permission to configure content types finder settings; form and
permission to configure the settings, index backend maybe other things, of the
finder itself. The index backend is the index not the finder note - it
is theoretically possible to end up with finders using the same index?

= Clear-up

Just that some other bits and bobs as the code came into sight.